Refactor countries search components
